+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 > Current Version: 0.1.1-alpha
 
 # 0.1.1-alpha
-- Accounts module
+- Feature: Accounts module
+- Feature: Module banner
 
 # 0.0.1-alpha / Jan 15th 2023
 - Project setup
-- Account module released
+- Feature: Account module released